refactor: 초기 엔티티 구조 리팩토링
