--- src/main/java/com/project/always/domain/member/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.project.always.domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> { // type, id
-
-}
