+++ src/main/java/com/project/always/utils/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.project.always.domain;
+package com.project.always.utils;
 
 import java.time.LocalDateTime;
 import javax.persistence.EntityListeners;
